// frontend/src/api/generated/models/MilestoneDto.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
export type MilestoneDto = {
    key: string;
    title: string;
    description: string;
    category: string;
    icon: string;
    threshold: number;
    currentValue: number;
    earned: boolean;
    earnedDate?: string | null;
    progressPercent: number;
};